Cleaned up commands and subsystems and renamed variablesrenamed
Removed extra print statements, changed the intake constants to just cube and cone speed, and negated them where needed in intakeSubsystem. in all commands, changed the instance name of IntakeSubsystem, to mIntakeSubsystem. Removed the extra lines of code in the execute command of IntakeCommand. Made setSpeed private, and used setSpeed in place of the current algorithms in intakeSubsystem methods.